Fix yt-dlp output file name

+++ share/gpodder/extensions/youtube-dl.py
@@ -85,6 +85,15 @@ class YoutubeCustomDownload(download.CustomDownload):
         self._reporthook = None
         self._prev_dl_bytes = 0
         self._episode = episode
+        self._partial_filename = None
+
+    @property
+    def partial_filename(self):
+        return self._partial_filename
+
+    @partial_filename.setter
+    def partial_filename(self, val):
+        self._partial_filename = val
 
     def retrieve_resume(self, tempname, reporthook=None):
         """
@@ -94,27 +103,47 @@ class YoutubeCustomDownload(download.CustomDownload):
         # outtmpl: use given tempname by DownloadTask
         # (escape % because outtmpl used as a string template by youtube-dl)
         outtmpl = tempname.replace('%', '%%')
-        res = self._ytdl.fetch_video(self._url, outtmpl, self._my_hook)
-        # Renaming is not required because the escaped percent is not escaped in the output file.
+        info, opts = self._ytdl.fetch_info(self._url, outtmpl, self._my_hook)
+        if program_name == 'yt-dlp':
+            self.partial_filename = os.path.join(
+                opts['paths']['home'], opts['outtmpl']['default']) % info
+        elif program_name == 'youtube-dl':
+            self.partial_filename = opts['outtmpl'] % info
+
+        res = self._ytdl.fetch_video(info, opts)
+        if program_name == 'yt-dlp':
+            # yt-dlp downloads to whatever file name it wants, so rename
+            filepath = res.get('requested_downloads', [{}])[0].get('filepath')
+            if filepath is None:
+                raise Exception("Could not determine youtube-dl output file")
+            if filepath != tempname:
+                logger.debug('yt-dlp downloaded to "%s" instead of "%s", moving',
+                             os.path.basename(filepath),
+                             os.path.basename(tempname))
+                os.remove(tempname)
+                os.rename(filepath, tempname)
+
         if 'duration' in res and res['duration']:
             self._episode.total_time = res['duration']
         headers = {}
         # youtube-dl doesn't return a content-type but an extension
         if 'ext' in res:
             dot_ext = '.{}'.format(res['ext'])
-            # See #673 when merging multiple formats, the extension is appended to the tempname
-            # by youtube-dl resulting in empty .partial file + .partial.mp4 exists
-            # and #796 .mkv is chosen by ytdl sometimes
-            for try_ext in (dot_ext, ".mp4", ".m4a", ".webm", ".mkv"):
-                tempname_with_ext = tempname + try_ext
-                if os.path.isfile(tempname_with_ext):
-                    logger.debug('youtube-dl downloaded to "%s" instead of "%s", moving',
-                                 os.path.basename(tempname_with_ext),
-                                 os.path.basename(tempname))
-                    os.remove(tempname)
-                    os.rename(tempname_with_ext, tempname)
-                    dot_ext = try_ext
-                    break
+            if program_name == 'youtube-dl':
+                # See #673 when merging multiple formats, the extension is appended to the tempname
+                # by youtube-dl resulting in empty .partial file + .partial.mp4 exists
+                # and #796 .mkv is chosen by ytdl sometimes
+                for try_ext in (dot_ext, ".mp4", ".m4a", ".webm", ".mkv"):
+                    tempname_with_ext = tempname + try_ext
+                    if os.path.isfile(tempname_with_ext):
+                        logger.debug('youtube-dl downloaded to "%s" instead of "%s", moving',
+                                     os.path.basename(tempname_with_ext),
+                                     os.path.basename(tempname))
+                        os.remove(tempname)
+                        os.rename(tempname_with_ext, tempname)
+                        dot_ext = try_ext
+                        break
+
             ext_filetype = util.mimetype_from_extension(dot_ext)
             if ext_filetype:
                 # YouTube weba formats have a webm extension and get a video/webm mime-type
@@ -299,9 +328,12 @@ class gPodderYoutubeDL(download.CustomDownloader):
             opts['format'] += '/' + fallback
         logger.debug('format=%s', opts['format'])
 
-    def fetch_video(self, url, tempname, reporthook):
+    def fetch_info(self, url, tempname, reporthook):
         opts = {
-            'outtmpl': tempname,
+            'paths': {'home': os.path.dirname(tempname)},
+            # Postprocessing in yt-dlp breaks without ext
+            'outtmpl': (os.path.basename(tempname) if program_name == 'yt-dlp'
+                        else tempname) + '.%(ext)s',
             'nopart': True,  # don't append .part (already .partial)
             'retries': 3,  # retry a few times
             'progress_hooks': [reporthook]  # to notify UI
@@ -309,7 +341,12 @@ class gPodderYoutubeDL(download.CustomDownloader):
         opts.update(self._ydl_opts)
         self.add_format(self.gpodder_config, opts)
         with youtube_dl.YoutubeDL(opts) as ydl:
-            return ydl.extract_info(url, download=True)
+            info = ydl.extract_info(url, download=False)
+            return info, opts
+
+    def fetch_video(self, info, opts):
+        with youtube_dl.YoutubeDL(opts) as ydl:
+            return ydl.process_video_result(info, download=True)
 
     def refresh_entries(self, ie_result):
         # only interested in video metadata
